+++ src/components/System/Debug/Debugging.ts
@@ -75,27 +75,27 @@ export function enableDebugging(): Promise<any> {
 
 // TODO: Comment this out for build time! Slows things down a lot.
 
-// /**
-//  * A function to activate debug mode (babylonjs)
-//  * @returns void
-//  */
-// function debugMode(): void {
-//     import(
-//         /* webpackChunkName: "debugLayer" */
-//         /* webpackMode: "lazy" */
-//         "@babylonjs/core/Debug/debugLayer"
-//     ).then(() => {
-//         return import(
-//             /* webpackChunkName: "inspector" */
-//             /* webpackMode: "lazy" */
-//             "@babylonjs/inspector"
-//         )
-//     }).then(() => {
-//         Vars.scene.debugLayer.show().then(() => {
-//             document.getElementById("inspector-host").style.zIndex = "15";
-//             document.getElementById("scene-explorer-host").style.zIndex = "15";
-//         });
-//     })
-// }
+/**
+ * A function to activate debug mode (babylonjs)
+ * @returns void
+ */
+function debugMode(): void {
+    import(
+        /* webpackChunkName: "debugLayer" */
+        /* webpackMode: "lazy" */
+        "@babylonjs/core/Debug/debugLayer"
+    ).then(() => {
+        return import(
+            /* webpackChunkName: "inspector" */
+            /* webpackMode: "lazy" */
+            "@babylonjs/inspector"
+        )
+    }).then(() => {
+        Vars.scene.debugLayer.show().then(() => {
+            document.getElementById("inspector-host").style.zIndex = "15";
+            document.getElementById("scene-explorer-host").style.zIndex = "15";
+        });
+    })
+}
 
-// window["debugMode"] = debugMode;
+window["debugMode"] = debugMode;