+++ src/components/UI/Vue/Components/OpenPopup/LoadAllVue.ts
@@ -5,6 +5,7 @@
 import { ModalComponent } from "./ModalComponent";
 import { SimpleModalComponent } from "./SimpleModalComponent";
 import { LoadSaveModalComponent } from "./LoadSaveModalComponent";
+import { MessagesComponent } from "../MessagesComponent";
 declare var Vue;
 
 /**
@@ -14,4 +15,5 @@ export function load(): void {
     new SimpleModalComponent().load(Vue);
     new LoadSaveModalComponent().load(Vue);
     new ModalComponent().load(Vue);
+    new MessagesComponent().load(Vue);
 }